lk: Use post-increment only in inffast.c
An old inffast.c optimization turns out to not be optimal anymore
with modern compilers, and furthermore was not compliant with the
C standard, for which decrementing a pointer before its allocated
memory is undefined. Per the recommendation of a security audit of
the zlib code by Trail of Bits and TrustInSoft, in support of the
Mozilla Foundation, this "optimization" was removed, in order to
avoid the possibility of undefined behavior.

Change-Id: I922c4d929854571585bcc5c3b64ced7283d85148
Git-commit: 9aaec95e82117c1cb0f9624264c3618fc380cecb
Git-repo: https://github.com/madler/zlib
Signed-off-by: Vishnu Teja <quic_vteja@quicinc.com>